pyproject.toml
[tool.poetry]
name = "photoshare"
version = "0.1.0"
description = ""
authors = ["Mizhik <vlad.mezhinskiy07@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
greenlet = "^3.0.3"
bcrypt = "4.0.1"
cloudinary = "^1.40.0"
uvicorn = "^0.30.3"
pydantic = {extras = ["email"], version = "^2.8.2"}
python-multipart = "^0.0.9"
qrcode = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
